+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## v0.1.19
+
+- Updated @xyflow/react to 12.3.6
+- Updated edges: make the name/label read-only, auto rename it when the source/target node is renamed
+- Dev dependencies update
+
 ## v0.1.18
 
 - Added react 19.0.0 support